requirements.txt
1numpy

opencv-contrib-python 
scikit-learn
matplotlib
tqdm

insightface
onnxruntime 
faiss-cpu